requirements.txt
asgiref==3.4.1
click==8.0.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.14
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
honcho==1.0.1
psycopg2-binary==2.8.6
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
